Fix type mis-match for struct.unpack
An error was occuring because `struct.unpack` was being called with
an `int` argument but it expects a `bytes` object. This was fixed
by converting the `int` arguments to `bytes` objects.

This was probably overlooked in the transition to python3.

Tested on a Watlow F4T device connecting through TCP:

```
python chamberconnectlibrary-test.py WatlowF4T TCP 192.168.0.222
```